refactor:(server): rename solution
