+++ .github/workflows/dotnet.yml
@@ -70,23 +70,23 @@ jobs:
 
       - name: 🧪 Run unit tests with coverage
         run: |
-          dotnet test ./test/MCP.FinnHub.Server.SSE.Tests.Unit/MCP.FinnHub.Server.SSE.Tests.Unit.csproj \
+          dotnet test ./test/FinnHub.MCP.Server.SSE.Tests.Unit/FinnHub.MCP.Server.SSE.Tests.Unit.csproj \
           --configuration Release \
           --collect:'XPlat Code Coverage' \
           --verbosity normal \
           -p:CollectCoverage=true \
           -p:CoverletOutputFormat=cobertura \
           -p:ExcludeByAttribute=GeneratedCodeAttribute \
-          -p:CoverletOutput=./test/MCP.FinnHub.Server.SSE.Tests.Unit/TestResults
+          -p:CoverletOutput=./test/FinnHub.MCP.Server.SSE.Tests.Unit/TestResults
 
       - name: 📂 Show test result files
-        run: ls -R ./test/MCP.FinnHub.Server.SSE.Tests.Unit/TestResults
+        run: ls -R ./test/FinnHub.MCP.Server.SSE.Tests.Unit/TestResults
 
       - name: 📤 Upload coverage to Codecov
         uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          files: ./test/MCP.FinnHub.Server.SSE.Tests.Unit/TestResults/coverage.cobertura.xml
+          files: ./test/FinnHub.MCP.Server.SSE.Tests.Unit/TestResults/coverage.cobertura.xml
           flags: unit-tests
           name: codecov-mcp-finnhub
 
@@ -94,4 +94,4 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
            name: test-results
-           path: ./test/MCP.FinnHub.Server.SSE.Tests.Unit/TestResults/
+           path: ./test/FinnHub.MCP.Server.SSE.Tests.Unit/TestResults/
