+++ finnhub-mcp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+#
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{D728C487-FA3C-4698-B853-4FB8041E6374}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{A7AF5572-5130-4729-9302-1DB158FB8A08}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FinnHub.MCP.Server.SSE", "src\FinnHub.MCP.Server.SSE\FinnHub.MCP.Server.SSE.csproj", "{859E69D9-3195-4DFC-A463-107739C456F1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FinnHub.MCP.Server.SSE.Tests.Unit", "tests\FinnHub.MCP.Server.SSE.Tests.Unit\FinnHub.MCP.Server.SSE.Tests.Unit.csproj", "{E12898E6-98E9-4BE7-8D0E-60E888D7FF55}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{859E69D9-3195-4DFC-A463-107739C456F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{859E69D9-3195-4DFC-A463-107739C456F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{859E69D9-3195-4DFC-A463-107739C456F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{859E69D9-3195-4DFC-A463-107739C456F1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E12898E6-98E9-4BE7-8D0E-60E888D7FF55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E12898E6-98E9-4BE7-8D0E-60E888D7FF55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E12898E6-98E9-4BE7-8D0E-60E888D7FF55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E12898E6-98E9-4BE7-8D0E-60E888D7FF55}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{859E69D9-3195-4DFC-A463-107739C456F1} = {D728C487-FA3C-4698-B853-4FB8041E6374}
+		{E12898E6-98E9-4BE7-8D0E-60E888D7FF55} = {A7AF5572-5130-4729-9302-1DB158FB8A08}
+	EndGlobalSection
+EndGlobal
