+++ src/FinnHub.MCP.Server.SSE/Application/ResultErrorType.cs
@@ -1,0 +1,21 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+//  <copyright>
+//    This file is part of FinnHub MCP Server and is licensed under the MIT License.
+//    See the LICENSE file in the project root for full license information.
+//  </copyright>
+//  <summary>
+//    // TODO Add summary
+//  </summary>
+// ---------------------------------------------------------------------------------------------------------------------
+
+namespace FinnHub.MCP.Server.SSE.Application;
+
+public enum ResultErrorType
+{
+    NotFound,
+    Unknown,
+    InvalidQuery,
+    ServiceUnavailable,
+    Timeout,
+    InvalidResponse
+}
