+++ src/FinnHub.MCP.Server.SSE/Properties/launchSettings.json
