+++ src/FinnHub.MCP.Server.SSE/appsettings.Development.json
