+++ src/FinnHub.MCP.Server.SSE/appsettings.json
