--- src/MCP.FinnHub.Server.SSE/Application/Features/Search/Services/ISearchService.cs
@@ -1,21 +1,0 @@
-﻿// ---------------------------------------------------------------------------------------------------------------------
-//  <copyright>
-//    This file is part of  FinnHub MCP project and is licensed under the MIT License.
-//    See the LICENSE file in the project root for full license information.
-//  </copyright>
-//  <summary>
-//    // TODO Add summary
-//  </summary>
-// ---------------------------------------------------------------------------------------------------------------------
-
-using MCP.FinnHub.Server.SSE.Application.Features.Search.Queries;
-using MCP.FinnHub.Server.SSE.Models;
-
-namespace MCP.FinnHub.Server.SSE.Application.Features.Search.Services;
-
-public interface ISearchService
-{
-    Task<Result<IReadOnlyList<StockSymbol>>> SearchSymbolsAsync(
-        SymbolSearchQuery query,
-        CancellationToken cancellationToken = default);
-}
