--- src/MCP.FinnHub.Server.SSE/GlobalUsings.cs
@@ -1,13 +1,0 @@
-﻿// ---------------------------------------------------------------------------------------------------------------------
-//  <copyright>
-//    This file is part of  FinnHub MCP project and is licensed under the MIT License.
-//    See the LICENSE file in the project root for full license information.
-//  </copyright>
-//  <summary>
-//    // TODO Add summary
-//  </summary>
-// ---------------------------------------------------------------------------------------------------------------------
-
-global using System.Text.Json;
-global using ModelContextProtocol.Protocol;
-global using ModelContextProtocol.Server;
