--- src/MCP.FinnHub.Server.SSE/MCP.FinnHub.Server.SSE.http
@@ -1,6 +1,0 @@
-@MCP.FinnHub.Server.SSE_HostAddress = http://localhost:5101
-
-GET {{MCP.FinnHub.Server.SSE_HostAddress}}/
-Accept: application/json
-
-###
