--- test/MCP.FinnHub.Server.SSE.Tests.Unit/MCP.FinnHub.Server.SSE.Tests.Unit.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-      <TargetFramework>net8.0</TargetFramework>
-      <ImplicitUsings>enable</ImplicitUsings>
-      <Nullable>enable</Nullable>
-      <IsPackable>false</IsPackable>
-      <IsTestProject>true</IsTestProject>
-      <RootNamespace>MCP.FinnHub.Server.SSE.Tests.Unit</RootNamespace>
-    </PropertyGroup>
-    <ItemGroup>       
-      <PackageReference Include="coverlet.collector" />
-      <PackageReference Include="Microsoft.NET.Test.Sdk" />
-      <PackageReference Include="xunit" />
-      <PackageReference Include="xunit.runner.visualstudio" />
-      <PackageReference Include="NSubstitute" />
-    </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\MCP.FinnHub.Server.SSE\MCP.FinnHub.Server.SSE.csproj" />
-  </ItemGroup>
-</Project>
