+++ tests/FinnHub.MCP.Server.SSE.Tests.Unit/FinnHub.MCP.Server.SSE.Tests.Unit.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+      <TargetFramework>net8.0</TargetFramework>
+      <ImplicitUsings>enable</ImplicitUsings>
+      <Nullable>enable</Nullable>
+      <IsPackable>false</IsPackable>
+      <IsTestProject>true</IsTestProject>
+      <RootNamespace>FinnHub.MCP.Server.SSE.Tests.Unit</RootNamespace>
+    </PropertyGroup>
+    <ItemGroup>       
+      <PackageReference Include="coverlet.collector" />
+      <PackageReference Include="Microsoft.NET.Test.Sdk" />
+      <PackageReference Include="xunit" />
+      <PackageReference Include="xunit.runner.visualstudio" />
+      <PackageReference Include="NSubstitute" />
+    </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\FinnHub.MCP.Server.SSE\FinnHub.MCP.Server.SSE.csproj" />
+  </ItemGroup>
+</Project>
